// src/app/model/trlawsuit.ts
export class trlawsuit {
    lawsuitId: number;
    documentNumber: string;
    debtCollectionNumber: string;
    idcard: string;
    propertyId: number;
    lawsuitType: string;
    lawyerName: string;
    capitalSuedPay: number;
    dateOfIntroduction: string;
    responsible: string;
    filingDate: string;
    InvestigationDate: string;
    court: string;
    courtDate: string;
    sequel: string;
    otherDetails: string;
    sentence: string;
    amountVerdict: number;
    lawyerFee: number;
    fee: number;
    litigationCosts: number;
    courtFees: number;
    amountOther: number;
    amountOtherNote: string;
    paymentAmount: number;
    status: string;
    active: string;
    remark: string;
    createdBy: string;
    createDate: Date;
    updatedBy: string;
    updateDate: Date;
    sourceId: string;
    source: string;
};